worked check_time trigger not to check section_id
It doesn't need to check for section id, the trigger needs to check that the professor (id = NEW.id), the mod(various logic), the semester and year match. This makes sure that the interim mod AA won't conflict with the non-interim mod A3 and so on. It also checks that the mod slot is NOT web, and thus, giving the ability to have multiple web courses AND make sure web slot doesn't conflict with any other slots